resource: changed TestResource value to s64, added serializer.

+++ src/togo/resource/types.hpp
@@ -156,7 +156,7 @@ struct ResourceMetadata {
 
 /// Test resource.
 struct TestResource {
-	u32 x;
+	s64 x;
 };
 
 // Forward declarations
